requirements.txt
aiohttp>=3.9.0
websockets>=12.0
numpy>=1.26.0
python-dotenv>=1.0.0
PyYAML>=6.0.1